+++ frontend/src/lib/components/circuit/block/BlockContentComponent.svelte
@@ -7,8 +7,15 @@
     import { isSkillItemRevealed } from "../../../logic/circuit/unlocked_skills.svelte";
     import type { SkillItem } from "../../../dto/circuit/edition/skill";
     import { getAdditionalIcons } from "../../../logic/preferences.svelte";
+    import {
+        type FocusModeBlockState,
+        FocusModeBlockStates,
+    } from "../../../data/focus_mode_block_state";
+    import { getFocusModeState } from "../../../logic/circuit/focusMode.svelte";
 
     let { block, completed }: { block: Block; completed: boolean } = $props();
+
+    let focusModeState: FocusModeBlockState = $derived(getFocusModeState(block.id));
 
     function isSkillItemVisible(item: SkillItem) {
         return item.column !== null && (!item.hidden || isSkillItemRevealed(item));
@@ -38,10 +45,10 @@
 </script>
 
 <div class="heading">
-    {#if block.blockType === "skill" && block.external}
+    {#if block.blockType === "skill" && block.external && focusModeState !== FocusModeBlockStates.DisabledInFocusMode}
         <span class="label">External</span>
     {/if}
-    {#if block.blockType === "skill" && !block.essential}
+    {#if block.blockType === "skill" && !block.essential && focusModeState !== FocusModeBlockStates.DisabledInFocusMode}
         <span class="label">Optional</span>
     {/if}
     <span class="name">
@@ -49,29 +56,32 @@
     </span>
 </div>
 
-{#if block.blockType === "skill"}
-    <TaskIconsComponent tasks={getItemsOnPath(block)}></TaskIconsComponent>
-{:else if hasEditorRights()}
-    <span>{block.items.length} {getLevel().items}</span>
-{:else}
-    {@const nrCompleted = getNumCompletedItems("essential")}
-    {@const nrTotal = getNumTotalItems("essential")}
-    {@const completedOpt = getNumCompletedItems("optional")}
-    {@const totalOpt = getNumTotalItems("optional")}
+<div
+    style={focusModeState === FocusModeBlockStates.DisabledInFocusMode ? "visibility: hidden" : ""}>
+    {#if block.blockType === "skill"}
+        <TaskIconsComponent tasks={getItemsOnPath(block)}></TaskIconsComponent>
+    {:else if hasEditorRights()}
+        <span>{block.items.length} {getLevel().items}</span>
+    {:else}
+        {@const nrCompleted = getNumCompletedItems("essential")}
+        {@const nrTotal = getNumTotalItems("essential")}
+        {@const completedOpt = getNumCompletedItems("optional")}
+        {@const totalOpt = getNumTotalItems("optional")}
 
-    <div class="completion-counters">
-        {#if nrTotal > 0}
-            <span>
-                {nrCompleted}/{nrTotal} completed
-            </span>
-        {/if}
-        {#if block.blockType === "submodule" && totalOpt > 0}
-            <span class="optional-counter">
-                {completedOpt}/{totalOpt} optional
-            </span>
-        {/if}
-    </div>
-{/if}
+        <div class="completion-counters">
+            {#if nrTotal > 0}
+                <span>
+                    {nrCompleted}/{nrTotal} completed
+                </span>
+            {/if}
+            {#if block.blockType === "submodule" && totalOpt > 0}
+                <span class="optional-counter">
+                    {completedOpt}/{totalOpt} optional
+                </span>
+            {/if}
+        </div>
+    {/if}
+</div>
 
 {#if completed && getAdditionalIcons()}
     <span class="checkmark fa-solid fa-check"></span>